Added the ability to specify 'required' parameters
Required parameters are a requested feature by Descriptor users. This
commit allows users to opt in to saying if a field is required or not.
This means required can be in three states: null, true, false. This was
done to support backward compatibility and not confuse implementations
in the wild.

#23